shift the viewer intrinsic to numba.unsafe
